+++ types/src/jsMain/kotlin/screeps/arena/api/ConstructionSite.kt
@@ -6,9 +6,9 @@ package screeps.arena.api
 /** A site of a structure which is currently under construction. */
 external open class ConstructionSite : GameObject {
     /** The current construction progress. */
-    val progress: Double?
+    val progress: Int?
     /** The total construction progress needed for the structure to be built. */
-    val progressTotal: Double?
+    val progressTotal: Int?
     /** The structure that will be built (when the construction site is completed). */
     val structure: Structure?
     /** Whether it is your construction site. */
